CPES/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cpes"
version = "0.1.0"
description = "Composable Persona Emulation Stack - A practical blueprint for building stable behavioral emulation"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "CPES Founder", email = "a.leilionas@vk.com"},
]
keywords = ["ai", "persona", "emulation", "cognitive", "llm", "behavioral"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "numpy>=1.21.0",
    "networkx>=2.6.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "loguru>=0.6.0",
    "pydantic>=1.8.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pytest-cov>=4.0.0",
]
swarms = [
    "swarms>=2.0.0",
]

[project.scripts]
cpes = "cpes.cli:main"

[project.urls]
Homepage = "https://github.com/cpes-team/cpes"
Repository = "https://github.com/cpes-team/cpes"
Documentation = "https://cpes.readthedocs.io"
"Bug Tracker" = "https://github.com/cpes-team/cpes/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cpes*"]

[tool.setuptools.package-data]
cpes = ["examples/*.py", "examples/*.yaml"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
